+++ src/server/handlers/participants/{Type}/{ID}.ts
@@ -38,8 +38,9 @@ export async function post(_context: Context, request: Request, h: ResponseToolk
     fspId: payload.fspId,
     subId: payload.partySubIdOrType
   }
-  await createPartyMapItem(partyMapItem)
-  return h.response().code(201)
+  const errResult = await createPartyMapItem(partyMapItem)
+  // istanbul ignore next
+  return h.response(errResult).code(!errResult ? 201 : errResult.statusCode)
 }
 
 export async function put(_context: Context, request: Request, h: ResponseToolkit): Promise<ResponseObject> {